// src/common/interfaces/signInInterface.ts
export interface IVisibilityPassword {
    password: boolean;
    confirmPassword: boolean;
}

// export interface ICHECKSCHEMA {
//     length: boolean;
//     dot: boolean;
//     number: boolean;
//     small: boolean;
//     special: boolean;
//     capital: boolean;
//     matched: boolean;
// }

export interface ISignInInitialValues {
    email: string;
    password: string;
    rememberMe:boolean;
}